requirements.txt
asgiref==3.9.1
attrs==25.3.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.7
dj-database-url==2.1.0
Django==5.1.12
django-cors-headers==4.4.0
django-debug-toolbar==6.0.0
django-defender==0.9.8
django-extensions==3.2.3
django-filter==25.1
django-oauth-toolkit==1.7.1
django-ratelimit==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.28.0
et_xmlfile==2.0.0
gunicorn==21.2.0
idna==3.10
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jwcrypto==1.5.6
numpy==2.3.2
oauthlib==3.3.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==6.4.0
referencing==0.36.2
requests==2.31.0
rpds-py==0.27.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.1.0
whitenoise==6.6.0
gunicorn==21.2.0
xhtml2pdf==0.2.17